Personal Projects
Daily and hourly averaging are projects that took in massive weather data CSV sets and outputs wind directions that were then used for writing a scientific poster. Raptor webscrape is a continuation of the same project: the aim is to automatically collect and process migrant raptor counts from different hawk watches in the united states.